+++ ui/litellm-dashboard/src/app/(dashboard)/users/_components/index.tsx
@@ -1,0 +1,1 @@
+export { default as ViewUserDashboard } from "./view_users";
